Updated Links and Moved Addresses and ABIs to config
