+++ src/hooks/data/useTokenOwner.jsx
@@ -6,9 +6,9 @@ import {
 
 import {
   ContractAbis,
-  ContractAddresses,
-  useContractCall
-} from '@/hooks/useContractCall'
+  ContractAddresses
+} from '@/config/contracts'
+import { useContractCall } from '@/hooks/useContractCall'
 
 const useTokenOwner = (tokenId) => {
   const [owner, setOwner] = useState('')
